+++ sha1.h
@@ -55,7 +55,7 @@ typedef struct
 	 * \brief Offset into \a data; points to the place where new data will be
 	 *  copied before it is processed.
 	 */
-	unsigned short offset;
+	uint32_t offset;
 } sha1Param;
 
 #ifdef __cplusplus
